+++ src/main/java/don/juan/matus/lib/bintree/GeneralCall.java
@@ -3,6 +3,7 @@ package don.juan.matus.lib.bintree;
 /**
  * Interface GeneralCall.
  */
+@FunctionalInterface
 public interface GeneralCall<T extends Comparable<T>> {
     void generalCall(Boolean theResult, Integer theCmp, BinTreeNodeInterface<T> resultNode, Object generalObject);
 }
