+++ src/main/java/don/juan/matus/lib/bintree/MergeStrategy.java
@@ -1,0 +1,8 @@
+package don.juan.matus.lib.bintree;
+
+@FunctionalInterface
+public interface MergeStrategy<T extends Comparable<T>> {
+
+    boolean solveIt(BinTreeNodeInterface<T> left, BinTreeNodeInterface<T> right);
+
+}
